% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_scan_range}
\alias{extract_scan_range}
\title{extract_scan_range}
\usage{
extract_scan_range(data_array, start_scan, end_scan)
}
\description{
function to extract a
particualr scan range of
the data array
}
