% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_openchrom_csv}
\alias{load_openchrom_csv}
\title{load_openchrom_csv}
\usage{
load_openchrom_csv(csv_file, uvdad = FALSE, range = NULL)
}
\description{
function to read openchrom csv
exported files directly into
mcrarvalidator array format
}
