% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_background}
\alias{remove_background}
\title{remove_background}
\usage{
remove_background(data_array)
}
\arguments{
\item{data_array}{mz channels in columns}
}
\description{
sweeps out the minimum of each column.
works on one and multiple chromatogram.
}
