% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_baseline}
\alias{remove_baseline}
\title{remove_baseline}
\usage{
remove_baseline(data_matrix)
}
\arguments{
\item{data_matrix}{contains signals along columns}
}
\description{
removes the baseline linear from first to last point
along columns, works on one and multiple chromatogram
}
